reactor to component styling and multiple files
