+++ scss/_global.scss
@@ -214,131 +214,6 @@ samp {
   height: 100%;
 
   // todo move these ones into their files
-  .account-list-sidebar {
-    --accountSidebarBg: #2c2c2c;
-
-    /* sidebar that moves in is z-index 100, so this higher */
-    z-index: 101;
-    background-color: var(--accountSidebarBg);
-
-    width: 58px;
-
-    flex-shrink: 0;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    --als-avatar-size: 42px;
-    --als-avatar-margin: 4px;
-    --als-active-indicator-color: white;
-
-    .account {
-      position: relative;
-
-      margin-top: var(--als-avatar-margin);
-      margin-bottom: var(--als-avatar-margin);
-      height: var(--als-avatar-size);
-      .avatar {
-        --local-avatar-size: var(--als-avatar-size);
-        --local-avatar-vertical-margin: 0px;
-        // --local-avatar-font-size: 26px;
-
-        div.content,
-        img.content {
-          border-radius: 10%;
-        }
-      }
-
-      &.active,
-      &:hover,
-      &.context-menu-active {
-        .avatar::before {
-          content: ' ';
-          position: absolute;
-          overflow: hidden;
-          background: var(--als-active-indicator-color);
-
-          --als-active-indicator-height: 25px;
-
-          top: calc(
-            calc(var(--als-avatar-size) / 2) - calc(
-                var(--als-active-indicator-height) / 2
-              )
-          );
-          margin-inline-start: -12px;
-          height: var(--als-active-indicator-height);
-          width: 8px;
-          border-radius: 20px;
-        }
-      }
-
-      &:hover {
-        .avatar::before {
-          --als-active-indicator-height: 15px;
-        }
-      }
-
-      &.context-menu-active {
-        .avatar::before {
-          --als-active-indicator-height: 10px;
-          opacity: 0.7;
-        }
-      }
-
-      .account-badge {
-        position: absolute;
-        z-index: 102;
-        bottom: -6px;
-        right: -5px;
-
-        .fresh-message-counter,
-        .bg-sync-disabled {
-          color: var(--unreadCountLabel);
-          background-color: var(--unreadCountBg);
-          text-align: center;
-          font-size: 9pt;
-          height: 18px;
-          padding: 0px 4pt;
-          line-height: 18px;
-          border-radius: 10px;
-          font-weight: bold;
-        }
-
-        .bg-sync-disabled {
-          background-color: grey;
-          padding: 0px 3pt;
-        }
-      }
-    }
-
-    .add-button {
-      margin-top: var(--als-avatar-margin);
-      margin-bottom: var(--als-avatar-margin);
-      min-height: var(--als-avatar-size);
-      min-width: var(--als-avatar-size);
-      background: none;
-
-      --add-button-color: #808080;
-
-      color: var(--add-button-color);
-
-      border-radius: 10%;
-      border: 1.3pt solid var(--add-button-color);
-      font-size: 22px;
-      &:hover {
-        color: var(--accountSidebarBg);
-        background: var(--add-button-color);
-      }
-    }
-
-    overflow-y: scroll;
-    scrollbar-width: 3px;
-    &::-webkit-scrollbar {
-      width: 0px;
-    }
-  }
-
   .login-screen {
     flex-grow: 1;
   }
